add manager package
