+++ manager/manager.go
@@ -1,0 +1,48 @@
+package manager
+
+import (
+	"errors"
+        "log"
+	"sync"
+	"time"
+
+	"github.com/m-lab/annotation-service/api"
+	"github.com/m-lab/annotation-service/geolite2"
+)
+
+var (
+	// ErrNilDataset is returned when CurrentAnnotator is nil.
+	ErrNilDataset = errors.New("CurrentAnnotator is nil")
+
+	// A mutex to make sure that we are not reading from the CurrentAnnotator
+	// pointer while trying to update it
+	currentDataMutex = &sync.RWMutex{}
+
+	// CurrentAnnotator points to a GeoDataset struct containing the absolute
+	// latest data for the annotator to search and reply with
+	CurrentAnnotator api.Annotator
+)
+
+func GetAnnotator(date time.Time) api.Annotator {
+	// TODO - use the requested date
+	// dateString := strconv.FormatInt(date.Unix(), encodingBase)
+	currentDataMutex.RLock()
+	ann := CurrentAnnotator
+	currentDataMutex.RUnlock()
+	return ann
+}
+
+// PopulateLatestData will search to the latest Geolite2 files
+// available in GCS and will use them to create a new GeoDataset which
+// it will place into the global scope as the latest version. It will
+// do so safely with use of the currentDataMutex RW mutex. It it
+// encounters an error, it will halt the program.
+func PopulateLatestData() {
+	data, err := geolite2.LoadLatestGeolite2File()
+	if err != nil {
+		log.Fatal(err)
+	}
+	currentDataMutex.Lock()
+	CurrentAnnotator = data
+	currentDataMutex.Unlock()
+}
